Changed the formula surrounding burn rate.  It used to be (1 / IgniteBurnRate), but that gave incorrect DPS numbers.  Changed it to be 1 + (1 - IgniteBurnRate).
As an example, if you have a Burn Rate modifier of 85% faster, your IgniteBurnRate will be 0.15.  2 - 0.15 = 1.85, which will then be the burn rate mod used in DPS calculations.

I may have missed an edge case, but I believe this will fix the DPS numbers on the Ignite section of the Calcs tab.