Use SHIFT to hide passive tree tooltips.
I wanted to see the suggested path when hovering unallocated
nodes, and the tooltips make it difficult.  This change makes
use of the `SHIFT` key which is already in similar use on
jewel nodes.  I tried it out when drawing a path, another
application of `SHIFT` in that context, and I was pleasantly
surprised it let me see what I was drawing out much easier.

There may be some other use of `SHIFT` that makes this change
conflict with a different use pattern.  If so, my second choice
approach might be to have a checkbox at bottom, alongside
`Show Node Power`, which could be labelled "Hide Tooltips."