Fix #1858 - crash when adding a modifier to a jewel
This is the same issue as was fixed in abyss jewels in 9d42348. If there
are no options in the crafting table, the "crafting table" option needs
to be hidden, or there will be an error when trying to create the
tooltip.

For good measure, this patch also prevents the crash by checking that
there's something in listMod before dereferencing it.